Added lightweight store abstraction
This store abstraction takes functions which are performed against the
wrapped store connection. When transactions are supported by the store,
each operation is executed in a transactional context. If there is no
support, just the operation is executed. The JenaTDBStore is an
implementation of this pattern. It supports read and write transactions.
This fact was also abstracted (cf. Lock and TransactionSupport).

Note: The abstraction itself is not complete and leaking is part of the
design.