Added AbstractRepository
AbstractRepository is a base implementation for all other repositories.
It relies on the Transformation class to create RDF from a Java
instance. By default, the label and id are translated to RDF. The
transformation can be adapted by overriding a protected method. Each
model derived from a Java instance is stored in an own named graph. Each
time an entity is saved, this named graph is overwritten.