Package structure Refactoring
The current package structure had interdependencies:
* The domain package was dependent on the persistence package and vice
versa. 
* The domain.repository service was part of the domain but actually
belongs to the infrastructure and, thus, the content was moved to new
package persistence.local. 
* Repository, Entity and AbstractEntity were part of the persistence
package but rather belong to the domain package. 
* Types in the domain package were moved to sub-package domain.model.
New package RepositoryService was renamed to PersistenceService. An
interface was extracted and the current implementation was moved to
persistence.local. 
* Importers were moved from io.impl to io.file. 
* Test package was removed.