Changed package structure
The package structure was changed in order to fit with the new SPABuilder
and the new implementation style. This means the SPA itself is no longer bound
to a specific configuration but just provides the necssary methods.
If a local or remote persistence service is used is part of the SPABuilder
configuration.