Added Virtuoso-Jena-Provider and extended depending builders/factories
The virtuoso-jena-provider is used to connect to a runnign virtuoso
instance. The SPABuilder now provides methods to create an instance with
a configured virtuoso connection. In addition the lower level factories
were adapted, too.
The test case covers the import of schemes and related data. Both plus
the corresponding data model are saved using the jena-virtuoso provider.
Then a query is sent to the SPARQL service and its results are
evaluated.