Hotfix: Remove all Datapools from a project
In the current implementation a ConcurrentModifactionError is thrown, when all
data pools are removed from the Project. The reason is, that the keys method
is called which returns a view on the underlying datapools map. The keys
are iterated and passed to the map for removal. This yields the concurrent
modification error. This change resolves this issue by first capturing all
datapools in a seperate list and, then, calling clear on the map.