Added detailed exception for XSD Import if the namespaces do not match
If a XSD/XML file-pair is imported and the base-namespaces do not match
then a NullPointer is thrown. This however does not make clear why the
NullPointer occurred. The message now provides possible reasons for the
failure. In addition the XES-XSD uses the default xes-standard.org
base-namespace.

related: https://trello.com/c/NWIFY3a3