findAll returns only real entities
The findAll method assumed, that only named graphs containing an
instance of an entity have a name which starts with the corresponding
RDF class. However, for PartialDataStores this assumption is not true,
because the corresponding data is stored in a graph which has the
entities name as id.

This change alters the behavior of the findAll method to correctly
identify graphs of instances to have trailing string graph.