Repaired breaking build due to container rule
The build was broken because of the VirtuosContainer ClassRule in SpaIntegrationTest. The code in the static initialiazer falsely assumed that the actual Docker container is already started when the VirtuosoContainer object is created. However, this is not the case. The container itself is started when the starting method of the VirtuosoContainer is invoked as ClassRule. Thus, the jdbc url was not yet available when requested in the static initializer. This commit extracts the Virtuoso related code into an own Integration test and checks whether the actual docker container is running before doing the tests. Further, some typos were corrected.